added StegoAlgorithm file, make some public function as internal
